Move tooltip adders from utils/tooltip package to skyblock/item/tooltip/adders
