+++ src/main/java/de/hysky/skyblocker/skyblock/item/tooltip/adders/ObtainedTooltip.java
@@ -1,4 +1,4 @@
-package de.hysky.skyblocker.utils.tooltip;
+package de.hysky.skyblocker.skyblock.item.tooltip.adders;
 
 import de.hysky.skyblocker.skyblock.item.tooltip.TooltipInfoType;
 import de.hysky.skyblocker.utils.ItemUtils;
@@ -20,7 +20,7 @@ public class ObtainedTooltip extends TooltipAdder {
 	private static final DateTimeFormatter OBTAINED_DATE_FORMATTER = DateTimeFormatter.ofPattern("MMMM d, yyyy").withZone(ZoneId.systemDefault()).localizedBy(Locale.ENGLISH);
 	private static final DateTimeFormatter OLD_OBTAINED_DATE_FORMAT = DateTimeFormatter.ofPattern("M/d/yy h:m a").withZone(ZoneId.of("UTC")).localizedBy(Locale.ENGLISH);
 
-	protected ObtainedTooltip(int priority) {
+	public ObtainedTooltip(int priority) {
 		super(priority);
 	}
